MVVM
- Creation of the model, view and viewModel directories
- Creation of the FetchCharacters Class with the function with the same
  name.
- All the data classes in the model directorie.
- The MainActivity in the view directorie.
