+++ app/src/main/java/com/example/rickandmortyapp/viewmodel/CharacterDiffCallback.kt
@@ -1,6 +1,7 @@
-package com.example.rickandmortyapp
+package com.example.rickandmortyapp.viewmodel
 
 import androidx.recyclerview.widget.DiffUtil
+import com.example.rickandmortyapp.model.Character
 
 class CharacterDiffCallback : DiffUtil.ItemCallback<Character>() {
     override fun areItemsTheSame(oldItem: Character, newItem: Character): Boolean {
