+++ app/src/main/java/com/example/rickandmortyapp/viewmodel/NetworkUtils.kt
@@ -1,4 +1,4 @@
-package com.example.rickandmortyapp
+package com.example.rickandmortyapp.viewmodel
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
